requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.2
opencv-contrib-python==4.2.0.32
packaging==20.3
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
Django==2.1.2
django-cors-headers==2.2.0
django-crontab==0.7.1
django-filter==2.0.0
djangorestframework==3.5.4
PyYAML==5.1.2
selenium==3.14.0
uiautomation==2.0.3
uiautomator==0.3.6
pywin32==227
pywinauto==0.6.3
PyAutoGUI==0.9.48
